Introduce CategorizedItem.
We refactor to be more OO. Feels like GuildedRose is turning into a
"Service Class" that knows everything. Rather we want it to just
understand how to categorise the items we stock. We create a new class
(to work around the restrictions of the Kata) which composes Item and
ItemCategory and now has the responsibility and the data needed to age
itself.

I'm not super comfortable with all this mutation... but we work within
the spirit of the Kata. To that end, the item attributes are copied into
then back out of the CategorizedItem. It's more code, but feels safer to
me.